Fix Initialize DB to include Categories table
The database actually initializes properly now. To do: set up
new templates, remove these sneaky .pyc files, eventually allow
multiple posts to be deleted at once in interface. I've got a
general layout that looks decent, so I need to get that going.
Also, I need to get the posts table to join to the category table
and to the users table, which is decently urgent.