Basic Category Functionality Added
Just committing this so I can go to bed. Categories can now be added, renamed,
and deleted. You can also view all posts in a certain category, and they're
also listed in the navigation expressly for that purpose. I still need to add
the ability to edit a category from its /category/view/ page. I also need to
add a label for the checkbox that allows you to delete a category. I need to
make a dropdown menu to change a post's category in /post/edit/. And of course,
I still have to finish whatever I haven't completed from the last commit.