Outlined todos in tables.py
I think the best way to do these joins is to make foreign keys. I know
I'm going to get into a mess. When I fix this, I'm also going to have to
take a look at /scripts/initializedb.py to make sure it makes the right
tables and everything.